pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "token-difr"
version = "0.1.2"
description = "Verify LLM outputs using Gumbel-Max sampling verification"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["llm", "verification", "gumbel-max", "sampling"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "torchaudio>=2.9.1",
    "datasets",
    "tqdm>=4.0.0",
    "python-dotenv",
    "openai>=1.0.0",
    "transformers>=4.30.0",
    "modal>=0.64.0",
    "tiktoken>=0.7.0",
    "vllm>=0.6.0",
]

[project.optional-dependencies]
vllm = []
all = []
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "ipykernel==6.29.0"
]

[project.urls]
Homepage = "https://github.com/adamkarvonen/token-difr"
Repository = "https://github.com/adamkarvonen/token-difr"

[tool.hatch.build.targets.wheel]
packages = ["src/token_difr"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.pip]
torch-backend = "cu128"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
